Step in the right direction, I think.

+++ RecipeBookApi/Services/DynamoRecipeService.cs
@@ -2,6 +2,7 @@
 using Common.Dynamo.Models;
 using RecipeBookApi.Models;
 using RecipeBookApi.Services.Contracts;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
@@ -43,7 +44,7 @@ namespace RecipeBookApi.Services
             return CreateRecipeViewModel(recipe, owner);
         }
 
-        public async Task<string> Create(RecipePostPutModel model)
+        public async Task<string> Create(RecipePostPutModel model, string executedById)
         {
             var newRecipe = new Recipe
             {
@@ -53,12 +54,17 @@ namespace RecipeBookApi.Services
                 Name = model.Name
             };
 
-            return await _recipeStorage.Create(newRecipe, model.ExecutedById);
+            return await _recipeStorage.Create(newRecipe, executedById);
         }
 
-        public async Task Update(string id, RecipePostPutModel model)
+        public async Task Update(string id, RecipePostPutModel model, string executedById)
         {
             var originalRecipe = await _recipeStorage.Read(id);
+            if (originalRecipe.CreatedById != executedById)
+            {
+                throw new Exception("You cannot update someone else's recipe");
+            }
+
             var updatedRecipe = new Recipe
             {
                 Description = model.Description,
@@ -67,11 +73,17 @@ namespace RecipeBookApi.Services
                 Name = model.Name
             };
 
-            await _recipeStorage.Update(originalRecipe, updatedRecipe, id, model.ExecutedById);
+            await _recipeStorage.Update(originalRecipe, updatedRecipe, id, executedById);
         }
 
-        public async Task Delete(string id)
+        public async Task Delete(string id, string executedById)
         {
+            var recipe = await _recipeStorage.Read(id);
+            if (recipe.CreatedById != executedById)
+            {
+                throw new Exception("You cannot delete someone else's recipe");
+            }
+
             await _recipeStorage.Delete(id);
         }
